Make values in light_manager_state_t unsigned and not short.
(a) It makes the types consistent with the API.
(b) We don't need to save those four bytes.

+++ firmware/main/light_manager.h
@@ -4,8 +4,8 @@
 #include <esp_event.h>
 
 typedef struct {
-	short int warm_value;
-	short int cold_value;
+	unsigned warm_value;
+	unsigned cold_value;
 } light_manager_state_t;
 
 ESP_EVENT_DECLARE_BASE(LIGHT_MANAGER_EVENT);
